requirements.txt
appdirs==1.4.3
attrs==19.3.0
bcrypt==3.1.7
Cerberus==1.3.2
cffi==1.14.0
click==7.1.2
Flask==1.1.2
flask-swagger-ui==3.25.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pathspec==0.8.0
pycparser==2.20
PyJWT==1.7.1
regex==2020.4.4
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
Werkzeug==1.0.1
